\name{DAMOCLES-internal}
\alias{compute_edgeTList}
\alias{DAMOCLES_loglik_rhs}
\alias{DAMOCLES_integrate_ODE}
\alias{DAMOCLES_loglik_choosepar}
\alias{DAMOCLES_simplex}
\alias{DAMOCLES_bin_trait_sim}
\alias{DAMOCLES_all_loglik}
\alias{DAMOCLES_all_loglik_rh}
\alias{DAMOCLES_all_loglik_old}
\alias{DAMOCLES_all_loglik_rhs}
\alias{DAMOCLES_all_integrate_ODE}
\alias{DAMOCLES_all_integrate_ODE_old}
\alias{DAMOCLES_all_loglik_choosepar}
\alias{DAMOCLES_all_ML}
\alias{DAMOCLES_all_M}
\alias{DAMOCLES_all_node}
\alias{DAMOCLES_mats}
\alias{DAMOCLES_r}
\alias{DAMOCLES_check_Mlist}
\alias{DAMOCLES_check_edgeTList}
\alias{HERACLES_find_trait_stait_combinations}
\alias{o.dectobin}
\alias{combsUse}
\alias{HERACLES_extractCI}
\alias{ses.mn_trait_d}
\alias{ses.m_trait_d}
\alias{mtd}
\alias{mntd}
\alias{HERACLES_speciation_sim}
\alias{HERACLES_assign_state}
\alias{integral_peak2}
\title{Internal DAMOCLES Functions}
\description{Internal DAMOCLES functions}
\details{These are not to be called by the user.}
\keyword{internal}
