% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMOCLES_sim_fast.R
\name{DAMOCLES_sim_fast}
\alias{DAMOCLES_sim_fast}
\title{Simulating DAMOCLES}
\usage{
DAMOCLES_sim_fast(
  phy,
  gamma_0,
  mu,
  root.state,
  keepExtinct = FALSE,
  stateOnly = FALSE
)
}
\arguments{
\item{phy}{phylogeny in phylo format}

\item{gamma_0}{initial per lineage rate of immigration (gamma)}

\item{mu}{per lineage rate of local extinction}

\item{root.state}{geographic state of ancestor i.e. present (1) or absent(0)}

\item{keepExtinct}{whether to retain data for extinct lineages}

\item{stateOnly}{whether to return only the presence or absence states of the extant species}
}
\value{
Either a table containing the following
columns: The first two columns contain the edge list of the phylogenetic tree, and the following two
contain when each edge starts and ends, and the fifth column if lineage is extant.
The sixth column contains the presence (1) or absence (0) of the species, which is identical
to the output if stateOnly == TRUE, when keepExtinct == FALSE and stateOnly == FALSE. Column
seven and eight contain information on respectively whetever the edge is a tip and its tip number.
}
\description{
Simulates DAMOCLES
}
\examples{

#create random phylogeny
library(ape)
phy = ape::rcoal(10)
		
#run DAMOCLES		
patable = DAMOCLES_sim_fast(
  phy,
  gamma_0 = 1.5,
  mu = 0,
  root.state = 1,
  keepExtinct = FALSE,
  stateOnly = FALSE
  )

#show presence/absence on the tree
patable$col = rep("black",dim(patable)[1])
patable$col[which(patable$state == 1)] = "red"
plot(phy,tip.col = patable$col)

}
\references{
Pigot, A.L. & R.S. Etienne (2015). A new dynamic null model for
phylogenetic community structure. Ecology Letters 18: 153-163.
}
\seealso{
\code{\link{DAMOCLES_sim}}
}
\author{
Bouwe R. Reijenga
}
\keyword{models}
