\name{DAMOCLES-internal}
\alias{DAMOCLES_loglik_rhs}
\alias{DAMOCLES_integrate_ODE}
\alias{compute_edgeTList}
\alias{DAMOCLES_loglik_choosepar}
\alias{DAMOCLES_simplex}
\title{Internal DAMOCLESD Functions}
\description{Internal DAMOCLES functions}
\details{These are not to be called by the user.}
\keyword{internal}
